Move files to fit with new, more grouped directory structure
